# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApexFinder Pro - Enterprise OSINT & HNWI Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: entityId
          in: query
          schema: { type: integer }
        - name: jurisdiction
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    patch:
      operationId: updateAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── RELATIONSHIPS ─────────────────────────────────────────────────────────
...
  /research/run:
    post:
      operationId: runResearch
      tags: [research]
      summary: Run MCTS autonomous research on an entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchRunInput"
      responses:
        "201":
          description: Research session created with MCTS path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSession"
...
components:
  schemas:
...
    Asset:
      type: object
      required: [id, category, identifier, jurisdiction, createdAt]
      properties:
        id: { type: integer }
        category: { type: string }
        identifier: { type: string }
        jurisdiction: { type: string }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        estimatedValue: { type: ["number", "null"] }
        address: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        lastActivityDate: { type: ["string", "null"] }
        sourceRegistry: { type: ["string", "null"] }
        ownerEntityId: { type: ["integer", "null"] }
        ownerName: { type: ["string", "null"] }
        createdAt: { type: string }
    AssetInput:
      type: object
      required: [category, identifier, jurisdiction]
      properties:
        category: { type: string }
        identifier: { type: string }
        jurisdiction: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        estimatedValue: { type: number }
        address: { type: string }
        description: { type: string }
        lastActivityDate: { type: string }
        sourceRegistry: { type: string }
        ownerEntityId: { type: integer }
    AssetUpdate:
      type: object
      properties:
        category: { type: string }
        identifier: { type: string }
        jurisdiction: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        estimatedValue: { type: number }
        address: { type: string }
        description: { type: string }
        lastActivityDate: { type: string }
        sourceRegistry: { type: string }
        ownerEntityId: { type: integer }

    # ── RELATIONSHIP ──────────────────────────────────────────────────────────
...
    ResearchSession:
      type: object
      required: [id, targetEntityId, crmStatus, createdAt]
      properties:
        id: { type: integer }
        targetEntityId: { type: integer }
        targetEntityName: { type: ["string", "null"] }
        winningPath: { type: ["string", "null"] }
        mctsSteps: { type: ["string", "null"] }
        generatedPitch: { type: ["string", "null"] }
        crmStatus: { type: string }
        lastContactDate: { type: ["string", "null"] }
        bayesianScoreAtRuntime: { type: ["number", "null"] }
        pathScore: { type: ["number", "null"] }
        realismClass: { type: ["string", "null"] }
        realismReason: { type: ["string", "null"] }
        researchBudget: { type: ["string", "null"] }
        createdAt: { type: string }
    ResearchRunInput:
      type: object
      required: [entityId]
      properties:
        entityId: { type: integer }
        depth: { type: integer }
        focusRegistries: { type: string }
...